requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
whisper
langchain
langchain-community
chromadb
sentence-transformers
huggingface-hub
PyPDF2
elevenlabs
requests
python-multipart